requirements.txt
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asyncmy==0.2.9
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
ecdsa==0.18.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.4
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.4
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
